+++ tools/cooja/apps/monitor/src/be/ac/umons/cooja/monitor/mon/backend/WarnSkipMon.java
@@ -41,12 +41,12 @@ public class WarnSkipMon extends SwitchableMon {
   }
   
   @Override
-  protected void skipState(int context, int entity, int state, MonTimestamp timestamp, double simTime, short nodeID) {
+  protected void skipState(int context, int entity, int state, MonTimestamp timestamp, long simTime, short nodeID) {
     logger.info(String.format("(mon) warning: state event skipped at (node: %d cpu: %f ms, %l cycles; sim: %f ms) ",
                                      timestamp.getMillis(), timestamp.getCycles(), simTime / 1000.));
   }
   @Override
-  protected void skipInfo(int context, int entity, byte[] info, MonTimestamp timestamp, double simTime, short nodeID) {
+  protected void skipInfo(int context, int entity, byte[] info, MonTimestamp timestamp, long simTime, short nodeID) {
     logger.info(String.format("(mon) warning: info event skipped at (node: %d cpu: %f ms, %l cycles; sim: %f ms)",
                                      timestamp.getMillis(), timestamp.getCycles(), simTime / 1000.));
   }
