+++ tools/cooja/apps/monitor/src/be/ac/umons/cooja/monitor/mon/switchable/StdMonBackend.java
@@ -59,7 +59,7 @@ public class StdMonBackend extends SwitchableMonBackend {
   }
 
   @Override
-  public void recordState(int context, int entity, int state, MonTimestamp timestamp, double simTime, short nodeID) throws MonException {
+  public void recordState(int context, int entity, int state, MonTimestamp timestamp, long simTime, short nodeID) throws MonException {
     /* Since we display directly on stdout we must take care of endianness and offset. */
     context = xtohs(context);
     entity  = xtohs(entity);
@@ -73,7 +73,7 @@ public class StdMonBackend extends SwitchableMonBackend {
   }
 
   @Override
-  public void recordInfo(int context, int entity, byte[] info, MonTimestamp timestamp, double simTime, short nodeID) throws MonException {
+  public void recordInfo(int context, int entity, byte[] info, MonTimestamp timestamp, long simTime, short nodeID) throws MonException {
     /* Since we display directly on stdout we must take care of endianness and offset. */
     context = xtohs(context);
     entity  = xtohs(entity);
